feat: 헤더 완성 및 불필요한 파일 제거
