+++ src/components/Header.tsx
@@ -13,10 +13,8 @@ export default function Header({ left, right }: { left: ReactNode; right?: React
 const headerCss = css({
   position: 'fixed',
   top: '0%',
-  left: '50%',
-  transform: 'translate(-50%, 0)',
   height: '64px',
-  width: '382px',
+  width: '430px',
   backgroundColor: 'black',
   padding: '24px',
   display: 'flex',
