--- src/page/ConfirmPage.tsx
@@ -1,20 +1,0 @@
-import { useNavigate } from 'react-router';
-import Header from '../components/Header';
-
-function ConfirmPage() {
-  const navigate = useNavigate();
-
-  return (
-    <>
-      <Header
-        left={
-          <button onClick={() => navigate(-1)}>
-            <img src="./assets/back.svg" />
-          </button>
-        }
-      />
-    </>
-  );
-}
-
-export default ConfirmPage;
